# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "technical-challenge",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/metrics": {
      "get": {
        "summary": "(internal) Get current process metrics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Prometheus metrics in text-based format",
            "content": {
              "text/plain": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "summary": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/tasks/{taskId}": {
      "get": {
        "summary": "",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "summary": "",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/tasks": {
      "get": {
        "summary": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "summary": "",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [],
        "responses": {}
      }
    },
    "/tasks/{taskId}/complete": {
      "put": {
        "summary": "",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    }
  }
}